feat: ensure noteResolvers succeed if token valid is provided

+++ src/graphql/resolvers/note.ts
@@ -1,19 +1,27 @@
-import mongoose, { mongo } from 'mongoose'
-import dayjs from 'dayjs'
 import NoteModel, { INote } from '../../database/models/note'
+import decodeToken, { JwtPayload } from '../../helpers/decodeToken'
+import mongoose from 'mongoose'
+import dayjs from 'dayjs'
 import { ApolloError } from 'apollo-server-micro'
+import { MicroRequest } from 'apollo-server-micro/dist/types'
+
+type IApolloContext = {
+    dbConnection: mongoose.Connection,
+    req: MicroRequest
+}
 
 export default {
     Query: {
         getAllNotes: async (
             parent,
             args,
-            { dbConnection }: { dbConnection: mongoose.Connection }
+            { dbConnection, req }: IApolloContext
         ): Promise<INote[]> => {
+            const jwtPayload = decodeToken(req)
             const Note: mongoose.Model<INote> = NoteModel(dbConnection)
             let list: INote[]
             try {
-                list = await Note.find().exec()
+                list = await Note.find({ userId: jwtPayload.id }).exec()
             } catch (error) {
                 console.error('> getAllNotes error: ', error)
                 throw new ApolloError('Error retrieving all notes')
@@ -24,12 +32,13 @@ export default {
         getNote: async (
             parent,
             { _id }: { _id: INote['_id'] },
-            { dbConnection }: { dbConnection: mongoose.Connection }
+            { dbConnection, req }: IApolloContext
         ): Promise<INote> => {
+            const jwtPayload = decodeToken(req)
             const Note: mongoose.Model<INote> = NoteModel(dbConnection)
 
             try {
-                const note = await Note.findById(_id).exec()
+                const note = await Note.findOne({ _id, userId: jwtPayload.id }).exec()
                 return note
             } catch (error) {
                 console.error('> getNote error: ', error);
@@ -40,15 +49,17 @@ export default {
     Mutation: {
         saveNote: async (
             parent,
-            { title, content }: { title: INote['title']; content: INote['content'] },
-            { dbConnection }: { dbConnection: mongoose.Connection }
+            { title, content }: INote,
+            { dbConnection, req }: IApolloContext
         ): Promise<INote> => {
+            const jwtPayload = decodeToken(req)
             const Note: mongoose.Model<INote> = NoteModel(dbConnection)
             try {
                 const note = await Note.create({
                     title,
                     content,
-                    date: dayjs().toDate()
+                    date: dayjs().toDate(),
+                    userId: jwtPayload.id
                 })
                 return note
             } catch (error) {
@@ -59,9 +70,17 @@ export default {
         deleteNote: async (
             parent,
             { _id }: { _id: INote['_id'] },
-            { dbConnection }: { dbConnection: mongoose.Connection }
+            { dbConnection, req }: IApolloContext
         ): Promise<INote> => {
+            const jwtPayload = decodeToken(req)
             const Note: mongoose.Model<INote> = NoteModel(dbConnection)
+
+            const noteFound = await Note.findById(_id).exec()
+
+            if(noteFound._id !== jwtPayload.id) {
+                throw new ApolloError('You are not allowed to execute this action')
+            }
+
             try {
                 const note = await Note.findByIdAndDelete(_id).exec()
                 return note
